backend/dataset-service/requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
click==8.1.7
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
motor==3.6.0
numpy==2.2.0
pandas==2.2.3
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
